requirements.txt
affine @ file:///home/conda/feedstock_root/build_artifacts/affine_1733762038348/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
branca @ file:///home/conda/feedstock_root/build_artifacts/branca_1734433375112/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1725267488082/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1746569525376/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560564262/work
cftime @ file:///home/conda/feedstock_root/build_artifacts/cftime_1725400458727/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1746214863626/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1746951630646/work
click-plugins @ file:///home/conda/feedstock_root/build_artifacts/click-plugins_1733731077999/work
cligj @ file:///home/conda/feedstock_root/build_artifacts/cligj_1733749956636/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1744743063425/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
folium @ file:///home/conda/feedstock_root/build_artifacts/folium_1740766619747/work
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1746913583525/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1743437245292/work
geopandas @ file:///home/conda/feedstock_root/build_artifacts/geopandas_1734346029138/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1745672166/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1746352627188/work
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1725459263097/work
mapclassify @ file:///home/conda/feedstock_root/build_artifacts/mapclassify_1733731066416/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
matplotlib==3.10.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
munkres==1.1.4
netCDF4 @ file:///home/conda/feedstock_root/build_artifacts/netcdf4_1745588555227/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1731521053/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1745118880429/work/dist/numpy-2.2.5-cp311-cp311-linux_x86_64.whl#sha256=07f2e2f5d9d4e980910d365ca5ef0485f92c6dfa75096d2efcd5823c6727ccf6
opencv-python==4.11.0
opencv-python-headless==4.11.0
orthority @ file:///home/conda/feedstock_root/build_artifacts/orthority_1746358008583/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
paho-mqtt @ file:///home/conda/feedstock_root/build_artifacts/paho-mqtt_1740051138625/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1744430449469/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1746646210358/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyogrio @ file:///home/conda/feedstock_root/build_artifacts/pyogrio_1746734539024/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1743089729650/work
pyproj @ file:///home/conda/feedstock_root/build_artifacts/pyproj_1742323228727/work
PySide6==6.8.3
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
rasterio @ file:///home/conda/feedstock_root/build_artifacts/rasterio_1742428572712/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rioxarray @ file:///home/conda/feedstock_root/build_artifacts/rioxarray_1745316908994/work
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1736496745968/work/dist/scikit_learn-1.6.1-cp311-cp311-linux_x86_64.whl#sha256=d3f0d01ef376cb8d55cead654a824455f1021124eef49bbd1bd03ba7954d3028
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1739790648951/work/dist/scipy-1.15.2-cp311-cp311-linux_x86_64.whl#sha256=034f74dff07afe60a41ed832856d3d7e1ec1374d115cd88d525ace252f5b36fc
shapely @ file:///home/conda/feedstock_root/build_artifacts/shapely_1743678311120/work
shiboken6==6.8.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
snuggs @ file:///home/conda/feedstock_root/build_artifacts/snuggs_1733818638588/work
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615904614/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1736692517942/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1744323578849/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
xarray @ file:///home/conda/feedstock_root/build_artifacts/xarray_1745980799320/work
xyzservices @ file:///home/conda/feedstock_root/build_artifacts/xyzservices_1745598038694/work
zstandard==0.23.0
